Clarify behavior of load statement
* loaded variables cannot be bound twice
* loaded variables cannot be reassigned
* loaded variables are not exported

These are the rules from the Java Starlark implementation.

Note "loaded variables cannot be reassigned" rule is somewhat
inconsistent: loaded variables behave the same as builtins: they
are available to the current module but not exported.

But builtins can be reassigned.

This might need to be changed for greater consistency, but also for
practical reasons: reexports.

Suppose there's a module which imports a function `foo` and exports
it. Currently it can be written like this:

```
load("module.star", _foo="foo")
foo = _foo # reexport
```

It is desirable to be able to write this shorter, as:

```
load("module.star", "foo")
foo = foo  # reexport
```

This resolves issue #37.