Note that sorted(reverse) must be a boolean
Currently the spec says that if reverse is True, then the sort is reversed. But it doesn't say what happens if it's a truthy non-boolean (e.g. 1). Looking at Python, it allows either a boolean or a number (probably for reverse compat), but in the interest of a sensibly implemented subset, suggest we require this to be boolean.