Explicitly state that Starlark syntax is strict subset of Python (#258)
State that Starlark syntax (but not semantic) is a strict subset
of Python, and this property will be maintained in the future
iterations of Starlark.

Context: users have asked what tools they can use to work with
starlark sources to implement codemods, linters, code formatters,
syntax highlighters, documentation generators etc.

One possible option is to recommend Python tools when AST is needed,
for example, Python's builtin `ast.parse`. However, to make that
recommendation correct, we should guarantee that Starlark syntax
(but not semantics) will stay strict subset of Python: so each valid
Starlark program could be parsed as Python, and tools won't need
to be heaviy rewritten when Starlark language changes.

This PR automatically closes
https://github.com/bazelbuild/starlark/issues/111
as it contravenes this newly established requirement.