Fix conformance test suite for Java, and disable for Go and Rust
Our CI has been failing because we haven't migrated from WORKSPACE to
MODULE.bazel. This commit mostly fixes this; a follow-up will address
the unrelated spec generator test failure.

The conformance tests for the Go and Rust implementations are disabled
until we're able to migrate their setup logic to MODULE.bazel (#305).

For the Java implementation, rather than try to reproduce the needed
setup logic in MODULE.bazel, we simply vendor in the deploy jar of the
Java Starlark interpreter's REPL.

The old WORKSPACE file is deleted, and the associated environment
initiailization in setup_workspace.sh is commented out. The bazelci
config is updated to use Ubuntu 24.04 to get a newer JDK compatible with
vendored jar.

Also simplified the conformance test runner framework a little.